150. 정범호 / AI (미완성 bgm바뀌는 기능), GameManager 가장가까운 Door리턴 함수 추가, 플레이어와 닿고있는 Door리턴 함수 추가, Door Sound리턴함수 추가, 스테이지파일수정

+++ TWEngine/Include/Resource/FbxLoader.cpp
@@ -437,7 +437,7 @@ void CFbxLoader::LoadUV(FbxMesh * pMesh, PFBXMESHCONTAINER pContainer,
 
 		vUV = pUV->GetDirectArray().GetAt(iUVIndex);
 
-		int iY = vUV.mData[1];
+		int iY = (int)vUV.mData[1];
 		float fY = 1.0f - (float)(vUV.mData[1] - iY);
 
 		pContainer->vecUV[iControlIndex].x = (float)(vUV.mData[0]);
